Made a template for index page
