+++ src/templates/indexPage/indexPage.module.scss
@@ -1,5 +1,5 @@
 /* Very temporary css */
-@import '../layouts/variables';
+@import '../../layouts/variables';
 
 .homePage {
   margin: 0;
